backend/pyproject.toml
[project]
name = "superdashboard"
version = "2.0.0"
description = "Full-stack engineering dashboard with AI assistant and micro-portal plugin system"
readme = "../README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Chris Chan"}
]
keywords = ["dashboard", "fastapi", "ai", "plugins", "engineering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Web Framework
    "fastapi==0.128.0",
    "uvicorn[standard]==0.40.0",
    "python-multipart>=0.0.9",

    # Database & ORM
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",

    # Data Validation
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",

    # HTTP Client
    "httpx>=0.28.0",
    "httpcore>=1.0.0",

    # AI/ML
    "openai==2.14.0",

    # Configuration
    "python-dotenv==1.2.1",

    # Plugin Dependencies
    "twilio==9.9.0",              # WhatsApp plugin
    "feedparser>=6.0.11",         # RSS reader plugin
    "APScheduler>=3.10.4",        # Scheduling
    "beautifulsoup4>=4.12.0",     # HTML parsing
    "lxml>=5.0.0",                # XML parsing
    "dnspython>=2.4.0",           # DNS monitoring plugin
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

test = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.28.0",  # For testing async HTTP
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
